+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,81 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.146
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.145
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [ab6f01a]
+  - ai@6.0.0-beta.144
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [81e29ab]
+  - ai@6.0.0-beta.143
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [7169511]
+- Updated dependencies [bbdcb81]
+  - ai@6.0.0-beta.142
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b1405bf]
+  - ai@6.0.0-beta.141
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [7fdd89d]
+  - ai@6.0.0-beta.140
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3bd2689]
+  - ai@6.0.0-beta.139
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.138
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [9ba4324]
+  - ai@6.0.0-beta.137
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3d83f38]
+  - ai@6.0.0-beta.136
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [afe7093]
   - ai@6.0.0-beta.135